src/python-jwt-server/requirements.txt
flask
pyjwt
firebase-admin
python-dotenv
requests
gunicorn
